fixed and refactored accounts sets screen
